requirements.txt
absl-py==0.15.0
accelerate==0.34.2
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aistudio-sdk==0.2.6
albucore==0.0.13
albumentations==1.4.10
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.4
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrdict3==2.0.2
attrs==24.2.0
autocommand==2.2.2
babel==2.16.0
backcall==0.2.0
backports.tarfile==1.2.0
bce-python-sdk==0.9.19
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cn2an==0.5.22
colorama==0.4.6
colorlog==6.8.2
comm==0.2.2
contourpy==1.1.1
crcmod==1.7
cryptography==43.0.0
cssselect==1.2.0
cssutils==2.11.1
cx-Oracle==8.3.0
cycler==0.12.1
Cython==3.0.10
datasets==2.21.0
debugpy==1.8.5
decorator==5.1.1
decord==0.6.0
dill==0.3.4
easydict==1.13
easyofd==0.3.5
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
faiss-cpu==1.8.0.post1
fast-tokenizer-python==1.0.2
fastapi==0.112.1
fasttext==0.9.3
ffmpy==0.4.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
flatbuffers==1.12
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gast==0.3.3
google-auth==2.34.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gradio==4.42.0
gradio_client==1.3.0
graphviz==0.20.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==8.2.0
importlib_resources==6.4.0
inflect==7.3.1
ipykernel==6.29.5
ipython==8.12.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jaraco.text==3.12.1
jedi==0.19.1
jeepney==0.8.0
jieba==0.42.1
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
# keras-nightly==2.5.0.dev2021032900
keyring==25.3.0
kiwisolver==1.4.5
lazy_loader==0.4
lmdb==1.5.1
loguru==0.7.2
lxml==5.2.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mindmap==0.0.65
mock==5.1.0
modelscope==1.17.1
more-itertools==10.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.12.2
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.16.2
opencv-contrib-python==4.10.0.84
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.7
oss2==2.18.6
packaging==24.1
paddle2onnx==1.2.7
paddleclas==2.5.2
paddlefsl==1.1.0
paddlenlp==2.8.1
paddleocr==2.8.1
pandas==2.0.3
parso==0.8.4
patsy==0.5.6
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
platformdirs==4.2.2
PPOCRLabel==2.1.3
premailer==3.10.0
prettytable==3.11.0
proces==0.1.7
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.4
pyclipper==1.3.0.post5
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.11
PyMuPDFb==1.24.9
PyMySQL==1.1.1
pypandoc==1.13
pyparsing==3.1.2
pypdfium2==4.30.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.4.1
pyxdg==0.28
PyYAML==6.0.2
pyzmq==26.1.1
rapidfuzz==3.9.6
rarfile==4.2
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
ruff==0.6.3
safetensors==0.4.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==3.3.3
semantic-version==2.10.0
sentencepiece==0.2.0
seqeval==1.2.2
setuptools==72.1.0
shapely==2.0.5
shellingham==1.5.4
simplejson==3.19.3
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.2
statsmodels==0.14.1
sympy==1.13.2
termcolor==1.1.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.12
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
tool-helpers==0.1.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trash-cli==0.24.5.26
triton==3.0.0
typeguard==4.3.0
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
visualdl==2.5.3
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.12.1
xformers==0.0.28.post1
xlrd==1.2.0
XlsxWriter==3.2.0
xmltodict==0.13.0
xxhash==3.5.0
yacs==0.1.8
yarl==1.9.4
